Avoid initializing nested Git repository
Previously a Git repository was initialized if a Cargo workspace was
detected. However, it's also possible for users to initialize rustlings
in an existing Git repository that doesn't contain a Cargo workspace.
In that case, it's still undesirable to initialize a nested Git
repository for rustlings. We therefore call Git directly to determine
if rustlings is being initialized in an existing Git repository.